Further split out integration tests
